Add files via upload
Please give it a try! I think I can do the other fields we decided, but it is an arduous task to create on these javascript files. I still have to do the planner implementation. 
I have some ideas, but we can talk about this when we are together. Things in how we defined the database and how to do foreign keys is a hard thing